Level 4: two tasks completed. Registration button added, and Username, Pasword, Display Name, and Alignment have been added to the login/registration page. As of know, when trying to login, the alignment needs to be specified as well, but it is not much of an issue since the alignment can be specified later by using the settings page. This is something that will be evaluated (or fixed tomorrow) as it is 3:50am right now :).
Item NONE to the alignment array has also been removed even though it appears as tho it has been added. I assume this change will appear in the next push, as this is a change that was added and then removed between the previous push and this one.

That is all for now.