Level 5: Completed! (with multiple-users issue resolved). Upon registration, a copy of all quests is made for the new user. This is the copy that the user later accesses and modifies. Also, a location of origin was added to the registration page to avoid reading null from Parse later in QuestListActivity. Some unnecessary imports were removed, and a minor bug fixed that occurred when returning to QuestListActivity (when onResume was called, the quest list would show the available quests even though the spinner was not set to "AVAILABLE"
On the other hand, perhaps there is a better way to duplicate a ParseObject (there is a method call it is supposed to inherit from its parent class, Object, yet it would not find such a method in any of my ParseObjects. I am inclined to believe I made a mistake that would not allow me to do that, which is why the duplication seems so extensive. On the positive side, it works, but may be not as smoothly as it could.

A possible addition to QuestListActivity is to add a pull quests request to the the menu. This is in case there are more quests added in the future, but one would have to be careful since pulling all quests would overwrite the progress the user has made for the quests that were pulled at registration.

That is all for now,
GZ