Now the user can get more quests (if there are any) from Parse. The option Get More Quests! was added to the menu in QuestListActivity. When this option is pressed, the code checks if the number of rows in QuestClass is the same as in QuestClass_UserObjectId, where QuestClass is the main copy of the quests and QuestClass_UserObjectId is the users copy, so if that is correct, then the user holds a copy of all the quests. On the other hand, if there are more rows in QuestClass than there are in QuestClass_UserObjectId, then the user can get some new quests, which it does at that point and saves QuestClass_UserObjectId in the background. It also updates the quest list shown to the user. Now the admin of the app can add new quests for the users and they can retrieve them.
Note: This relies heavily on the assumption that the admin will not delete quests from QuestClass, which is a valid assumption since it is a quest game after all and users want more things to do in the game, and not less.

This is the last commit & push before 2:30PM (20/04/2014). Since it was not clear if there was a due time on the due date, the repo can be reverted to this commit & push.